Move Report templating methods to dedicated file
- report.go -> template.go
