+++ output/report.go
@@ -11,7 +11,6 @@ import (
 	"os"
 	"strconv"
 	"strings"
-	"text/template"
 	"time"
 
 	"github.com/benchttp/runner/ansi"
@@ -164,61 +163,6 @@ func (rep *Report) String() string {
 	b.WriteString(line("Mean response time", msString(mean)))
 	b.WriteString(line("Total duration", msString(bk.Duration)))
 	return b.String()
-}
-
-// applyTemplate applies Report to a template using given pattern and returns
-// the result as a string. If pattern == "", it returns errTemplateEmpty.
-// If an error occurs parsing the pattern or executing the template,
-// it returns errTemplateSyntax.
-func (rep *Report) applyTemplate(pattern string) (string, error) {
-	if pattern == "" {
-		return "", errTemplateEmpty
-	}
-
-	t, err := template.
-		New("report").
-		Funcs(rep.templateFuncs()).
-		Parse(pattern)
-	if err != nil {
-		return "", fmt.Errorf("%w: %s", errTemplateSyntax, err)
-	}
-
-	var b strings.Builder
-	if err := t.Execute(&b, rep); err != nil {
-		return "", fmt.Errorf("%w: %s", errTemplateSyntax, err)
-	}
-
-	return b.String(), nil
-}
-
-func (rep *Report) templateFuncs() template.FuncMap {
-	return template.FuncMap{
-		"stats": func() basicStats {
-			if rep.stats.isZero() {
-				rep.stats.Min, rep.stats.Max, rep.stats.Mean = rep.Benchmark.Stats()
-			}
-			return rep.stats
-		},
-
-		"event": func(rec requester.Record, name string) time.Duration {
-			for _, e := range rec.Events {
-				if e.Name == name {
-					return e.Time
-				}
-			}
-			return 0
-		},
-
-		"fail": func(a ...interface{}) string {
-			if rep.errTplFailTriggered == nil {
-				rep.errTplFailTriggered = fmt.Errorf(
-					"%w: %s",
-					ErrTplFailTriggered, fmt.Sprint(a...),
-				)
-			}
-			return ""
-		},
-	}
 }
 
 // HTTPRequest returns the *http.Request to be sent to Benchttp server.
